change back soustraction
change the way to substract backgroud to avoid 16 bits problems

+++ visu/visualElectrons.py
@@ -1105,7 +1105,7 @@ class SEEELECTRONS(QMainWindow):
                 self.winOpt.dataBgExist is True):
 
             try:
-                self.data = self.data-self.winOpt.dataBg
+                self.data = self.data.astype(np.int32) - self.winOpt.dataBg.astype(np.int32)
             except:
                 msg = QMessageBox()
                 msg.setIcon(QMessageBox.Icon.Critical)
